pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "novomolgen"
version = "0.1.0"
description = "Foundation models for de novo small-molecule generation."
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
    "openbabel-wheel==3.1.1.23",
    "openmm",
    "pdbfixer",
    "rdkit",
    "syba @ git+https://github.com/lich-uct/syba.git",
    "xtb",
    "crest",
    "lightgbm==4.3.0",
    "deepsmiles==1.0.1",
    "transformers>=5,<6",
    "accelerate>=1,<2",
    "datasets==2.20.0",
    "hydra-core==1.3",
    "hydra-colorlog",
    "hydra-optuna-sweeper",
    "hydra-joblib-launcher",
    "torchmetrics",
    "captum",
    "wandb==0.17.4",
    "molvs==0.1.1",
    "selfies==2.1.1",
    "PyTDC",
    "pandarallel",
    "grpcio",
    "flash-attn @ https://github.com/mjun0812/flash-attention-prebuild-wheels/releases/download/v0.7.16/flash_attn-2.7.4%2Bcu128torch2.10-cp310-cp310-linux_x86_64.whl",
    "pyarrow",
    "openpyxl",
    "h5py",
    "jupyterlab",
    "ipywidgets",
    "umap-learn",
    "xgboost==1.0.2",
    "torch",
    "torchvision",
    "safe-mol==0.1.13",
    "psutil==6.0.0",
    "setuptools<81",
    "py-cpuinfo==9.0.0",
    "numpy",
    "pandas",
    "scikit-learn==1.2.0",
    "scipy",
    "rootutils",
    "pre-commit",
    "rich",
    "pytest",
    "sh",
    "omegaconf",
    "tqdm",
    "pynvml==11.5.3",
    "pickleshare",
    "fire",
    "loguru",
    "sentencepiece==0.2.0",
    "trl",
    "tabulate==0.9.0",
    "pdb-tools",
    "packaging==24.1",
    "ninja",
    "wheel",
    "deepspeed==0.14.4",
    "torchdata==0.8.0",
    "multiprocess",
    "pyscf",
    "morfeus-ml",
    "geometric",
    "pyberny",
    "wurlitzer",
    "sqlalchemy",
    "medchem",
    "polars-u64-idx",
    "lightning",
    "evaluate",
    "descriptastorus",
    "pyscreener",
    "group-selfies @ git+https://github.com/aspuru-guzik-group/group-selfies.git",
    "atomInSmiles==1.0.2",
    "polanyi @ git+https://github.com/kjelljorner/polanyi.git",
    "geodesic-interpolate @ git+https://github.com/virtualzx-nad/geodesic-interpolate.git",
    "fcd_torch==1.0.7",
    "rd-filters @ git+https://github.com/PatWalters/rd_filters.git",
    "hf-transfer>=0.1.9",
]

[tool.setuptools]
py-modules = []

[tool.gantry]
workspace = "ai2/oe-adapt-code"
gh_token_secret = "michaeln_GITHUB_TOKEN"
budget = "ai2/oe-adapt"
log_level = "warning"
quiet = false

[tool.uv]
package = false


[tool.ruff]
lint.ignore = ["E501", "E741", "W605"]
lint.select = ["E", "F", "I", "W"]
line-length = 119

[lint.isort]
lines-after-imports = 2
known-first-party = ["src"]
